pyproject.toml
[project]
name = "fresnel-lighting"
version = "0.1.0"
description = "A chronobiology-powered lighting agent for Philips Hue"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "claude-agent-sdk",
    "openai>=2.29.0",
    "phue>=1.1",
    "python-dotenv",
    "rich",
]

[project.optional-dependencies]
calendar = [
    "google-api-python-client",
    "google-auth-oauthlib",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fresnel"]

[project.scripts]
fresnel = "fresnel.agent:main"
